Fix one additional migration test that was producing flaky results. Also suppress deletion errors in tests when the file has already been deleted.
PiperOrigin-RevId: 241722082

+++ ScienceJournalTests/Helpers/MetadataManager+Testing.swift
@@ -36,6 +36,10 @@ extension MetadataManager {
   }
 
   func deleteRootDirectory() {
+    guard FileManager.default.fileExists(atPath: rootURL.path) else {
+      return
+    }
+
     do {
       try FileManager.default.removeItem(at: rootURL)
     } catch {
